requirements.txt
Django==5.0.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-markdownx==4.0.2
markdown==3.5.1
bleach==6.1.0
python-decouple==3.8
